Refactored command design pattern
